Fix : recommendSearch 빌드오류 suspense 추가
